gateway/requirements.txt
# Core framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# PII detection (Microsoft Presidio)
presidio-analyzer>=2.2.0,<3.0.0
presidio-anonymizer>=2.2.0,<3.0.0
spacy>=3.8.0,<4.0.0

# Database
aiosqlite>=0.20.0,<1.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0

# WebSocket
websockets>=14.0,<15.0

# Configuration
pyyaml>=6.0.0,<7.0.0

# Utilities
python-multipart>=0.0.18
httpx>=0.28.0,<1.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
